requirements.txt
djangorestframework
django-rest-authtoken
djangorestframework-simplejwt
drf-yasg
django-rest-swagger
